refactor(api): compact and rename top-level module
