+++ docs/source/api/collector.rst
