+++ docs/source/api/device.rst
