+++ docs/source/api/libnvml.rst
